Support vsite in template matching
