+++ tests/test_frontend/test_atype_from_template.py
@@ -3,6 +3,7 @@ import openmm.unit as unit
 from dmff.api.topology import TopologyData
 from dmff.api.xmlio import XMLIO
 from dmff.operators.templatetype import TemplateOperator
+from dmff.api.vstools import addVSiteToTopology, TwoParticleAverageSite, ThreeParticleAverageSite, TemplateVSitePatcher, SMARTSVSitePatcher
 import json
 import pytest
 
@@ -39,7 +40,33 @@ def test_atype_from_template_multi():
     for na, a in enumerate(topdata.atom_meta):
         print(a)
 
+def test_atype_from_template_with_vsite():
+    pdb = app.PDBFile("tests/data/template_vsite.pdb")
+    top = pdb.topology
+    topdata = TopologyData(top)
+    vslist = []
+    patcher = TemplateVSitePatcher()
+    patcher.loadFile("tests/data/template_and_vsite.xml")
+    patcher.parse()
+    patcher.patch(topdata, None, vslist)
+    newtop, new_vslist = addVSiteToTopology(top, vslist)
+    topdata = TopologyData(newtop)
+    topdata.addVirtualSiteList(new_vslist)
+
+    for nres in range(newtop.getNumResidues()):
+        topdata.setOperatorToResidue(nres, "template")
+
+    io = XMLIO()
+    io.loadXML("tests/data/template_and_vsite.xml")
+    ret = io.parseXML()
+
+    operator = TemplateOperator(ret)
+    operator.operate(topdata)
+    for na, a in enumerate(topdata.atom_meta):
+        print(a)
+
 
 if __name__ == "__main__":
-    test_atype_from_template_single()
-    test_atype_from_template_multi()
+    # test_atype_from_template_single()
+    # test_atype_from_template_multi()
+    test_atype_from_template_with_vsite()
